llama2-binding/llamaServer/requirements.txt
websockets==11.0.3
gpt4all==1.0.8
python-dotenv==1.0.0
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
sacremoses==0.0.53
transformers==4.30.2
transformers[sentencepiece]==4.30.2
sentence-transformers==2.2.2